refactor(web): Relocate IconButton component file
Move IconButton component to a more specific file location and update import paths. We're moving away from index.ts files for the sake of more declarative imports. Why? Mostly because TK said so TBH.

https://tkdodo.eu/blog/please-stop-using-barrel-files
